Add modseq-based message update tracking
Introduce a "modification sequence number" similar to the one
defined in RFC 7162. This allows one to track mailbox updates which
happened between two polls.

Three counters are added to the messages table: created, updated,
and expunged.

A monotically increasing counter is used instead of the current time
to avoid running into issues related to time adjustments, multiple
operations happening at the same time, and synchronization across
multiple servers.

Right now the counter is global (an BIGINT should be enough for
hundreds of years), but it could be made per-mailbox in the future
if desired.