Add text to the continuation request response line
74a489458d5f updated imapproxy to also support literals in the "LOGIN"
command, but it sends a continuation request response of just "+".
However, the grammar in RFC 3501 §9 specifies that the continuation
request response needs to have a space and at least one character after
the "+":

	continue-req    = "+" SP (resp-text / base64) CRLF
	[...]
	resp-text       = ["[" resp-text-code "]" SP] text
	[...]
	text            = 1*TEXT-CHAR
	TEXT-CHAR       = <any CHAR except CR and LF>

This causes github.com/jgosmann/dmarc-metrics-exporter to throw an error
when logging in, so this isn't just a theoretical issue.

The "real" Dovecot IMAP server behind the proxy uses "+ OK" as the
continuation request response, so this commit updates imapproxy to do
the same.